fix: change price and credit type to decimal

+++ Models/SubscriptionPlansSchema.js
@@ -3,7 +3,7 @@ const subsPlanTable = `
     CREATE TABLE if not exists subscription_plan (
         name VARCHAR(255) PRIMARY KEY ,
         duration INTEGER NOT NULL,
-        price INTEGER  NOT NULL
+        price DECIMAL(15,2) NOT NULL
     );
 `
 module.exports = subsPlanTable;
